backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
ultralytics==8.1.0
transformers==4.36.0
torch==2.1.0
faiss-cpu==1.7.4
Pillow==10.1.0
numpy==1.26.2
